refactor to fix validation public health management fields, fixed broken front end tests post rebase, fixed system tests
