+++ app/javascript/components/enrollment/steps/Contact.js
@@ -400,12 +400,17 @@ class Contact extends React.Component {
               </Form.Row>
             </Form>
             {this.props.previous && !this.props.hidePreviousButton && (
-              <Button variant="outline-primary" size="lg" className="btn-square px-5" onClick={this.props.previous}>
+              <Button id="enrollment-previous-button" variant="outline-primary" size="lg" className="btn-square px-5" onClick={this.props.previous}>
                 Previous
               </Button>
             )}
             {this.props.next && (
-              <Button variant="outline-primary" size="lg" className="float-right btn-square px-5" onClick={() => this.validate(this.props.next)}>
+              <Button
+                id="enrollment-next-button"
+                variant="outline-primary"
+                size="lg"
+                className="float-right btn-square px-5"
+                onClick={() => this.validate(this.props.next)}>
                 Next
               </Button>
             )}
