+++ app/javascript/tests/mocks/mockPatients.js
@@ -897,7 +897,7 @@ const mockPatient4 = {
   source_of_report: null,
   source_of_report_specify: null,
   submission_token: 'ed3535055837367e9edb14b131b130b5f4c70e15',
-  symptom_onset: null,
+  symptom_onset: '2021-08-04',
   travel_related_notes: null,
   travel_to_affected_country_or_area: true,
   updated_at: '2020-09-13T20:32:02.000Z',
@@ -905,7 +905,7 @@ const mockPatient4 = {
   user_defined_id_cdc: '4677015425',
   user_defined_id_nndss: '38966610-6',
   user_defined_id_statelocal: 'EX-771721',
-  user_defined_symptom_onset: null,
+  user_defined_symptom_onset: true,
   was_in_health_care_facility_with_known_cases: true,
   was_in_health_care_facility_with_known_cases_facility_name: 'ABC Care',
   white: null,
